Move IconHandle::Priv implemantation away
